+++ ase/neb.py
@@ -5,6 +5,7 @@ from math import sqrt
 import numpy as np
 
 import ase.parallel as mpi
+from ase.build import minimize_rotation_and_translation
 from ase.calculators.calculator import Calculator
 from ase.calculators.singlepoint import SinglePointCalculator
 from ase.io import read
@@ -14,8 +15,18 @@ from ase.utils.geometry import find_mic
 
 class NEB:
     def __init__(self, images, k=0.1, climb=False, parallel=False,
+                 cornercutting=False, remove_rotation_and_translation=False,
                  world=None):
         """Nudged elastic band.
+        
+        Paper I:
+            
+            G. Henkelman and H. Jonsson, Chem. Phys, 113, 9978 (2000).
+            
+        Paper II:
+            
+            G. Henkelman, B. P. Uberuaga, and H. Jonsson, Chem. Phys,
+            113, 9901 (2000).
 
         images: list of Atoms objects
             Images defining path from initial to final state.
@@ -25,6 +36,13 @@ class NEB:
             Use a climbing image (default is no climbing image).
         parallel: bool
             Distribute images over processors.
+        cornercutting: bool
+            If True, springs are treated as real ones with forces in the
+            spring-directions (not the estimated tangent-directions).
+        remove_rotation_and_translation: bool
+            If True, the rotation and translation along the path will be
+            minimized.  Only useful for molecules without any constraints.
+            systems
         """
         self.images = images
         self.climb = climb
@@ -32,11 +50,14 @@ class NEB:
         self.natoms = len(images[0])
         self.nimages = len(images)
         self.emax = np.nan
-
+        
+        self.remove_rotation_and_translation = remove_rotation_and_translation
+        self.cornercutting = cornercutting
+        
         if isinstance(k, (float, int)):
             k = [k] * (self.nimages - 1)
         self.k = list(k)
-
+        
         if world is None:
             world = mpi.world
         self.world = world
@@ -45,8 +66,11 @@ class NEB:
             assert world.size == 1 or world.size % (self.nimages - 2) == 0
 
     def interpolate(self, method='linear', mic=False):
+        if self.remove_rotation_and_translation:
+            minimize_rotation_and_translation(self.images[0], self.images[-1])
+        
         interpolate(self.images, mic)
-
+                 
         if method == 'idpp':
             self.idpp_interpolate(traj=None, log=None, mic=mic)
 
@@ -85,25 +109,41 @@ class NEB:
                 image.get_calculator().set_atoms(image)
             except AttributeError:
                 pass
-
+    
     def get_forces(self):
         """Evaluate and return the forces."""
         images = self.images
         forces = np.empty(((self.nimages - 2), self.natoms, 3))
-        energies = np.empty(self.nimages - 2)
+        energies = np.empty(self.nimages)
+
+        if self.remove_rotation_and_translation:
+            # Remove translation and rotation between
+            # images before computing forces:
+            for i in range(1, self.nimages):
+                minimize_rotation_and_translation(images[i - 1], images[i])
+
+        # Get initial and final energy:
+        for i in [0, -1]:
+            try:
+                energies[i] = images[i].get_potential_energy()
+            except RuntimeError:
+                energies[i] = np.nan
 
         if not self.parallel:
             # Do all images - one at a time:
             for i in range(1, self.nimages - 1):
-                energies[i - 1] = images[i].get_potential_energy()
+                energies[i] = images[i].get_potential_energy()
                 forces[i - 1] = images[i].get_forces()
+            
         elif self.world.size == 1:
             def run(image, energies, forces):
+                """Target function for thread."""
                 energies[:] = image.get_potential_energy()
                 forces[:] = image.get_forces()
+                
             threads = [threading.Thread(target=run,
                                         args=(images[i],
-                                              energies[i - 1:i],
+                                              energies[i:i + 1],
                                               forces[i - 1:i]))
                        for i in range(1, self.nimages - 1)]
             for thread in threads:
@@ -114,7 +154,7 @@ class NEB:
             # Parallelize over images:
             i = self.world.rank * (self.nimages - 2) // self.world.size + 1
             try:
-                energies[i - 1] = images[i].get_potential_energy()
+                energies[i] = images[i].get_potential_energy()
                 forces[i - 1] = images[i].get_forces()
             except:
                 # Make sure other images also fail:
@@ -127,38 +167,78 @@ class NEB:
 
             for i in range(1, self.nimages - 1):
                 root = (i - 1) * self.world.size // (self.nimages - 2)
-                self.world.broadcast(energies[i - 1:i], root)
+                self.world.broadcast(energies[i:i + 1], root)
                 self.world.broadcast(forces[i - 1], root)
 
-        imax = 1 + np.argsort(energies)[-1]
-        self.emax = energies[imax - 1]
+        imax = 1 + np.argsort(energies[1:-1])[-1]
+        self.emax = energies[imax]
 
-        tangent1 = find_mic(images[1].get_positions() -
-                            images[0].get_positions(),
-                            images[0].get_cell(), images[0].pbc)[0]
+        def dist(i1, i2):
+            """Vector pointing from image i1 to i2."""
+            return find_mic(images[i2].positions - images[i1].positions,
+                            images[i1].cell, images[i1].pbc)[0]
+
+        if self.cornercutting:
+            # Find length of springs:
+            t = dist(0, -1)
+            length = np.linalg.norm(t) / (self.nimages - 1)
+            
         for i in range(1, self.nimages - 1):
-            tangent2 = find_mic(images[i + 1].get_positions() -
-                                images[i].get_positions(),
-                                images[i].get_cell(),
-                                images[i].pbc)[0]
-            if i < imax:
-                tangent = tangent2
-            elif i > imax:
-                tangent = tangent1
-            else:
-                tangent = tangent1 + tangent2
+            t1 = -dist(i, i - 1)
+            t2 = dist(i, i + 1)
+            nt1 = np.linalg.norm(t1)
+            nt2 = np.linalg.norm(t2)
 
-            tt = np.vdot(tangent, tangent)
+            # Tangents are bisections of spring-directions
+            # (formula 2 of paper I)
+            if (i == 1 and np.isnan(energies[0]) or
+                i == self.nimages - 2 and np.isnan(energies[-1])):
+                tangent = t1 / nt1 + t2 / nt2
+            else:
+                # Tangents are improved according to formulas 8, 9, 10,
+                # and 11 of paper I.
+                # First and last image in NEB cannot be improved since
+                # the energy is not known for the static points
+                if energies[i + 1] > energies[i] > energies[i - 1]:
+                    tangent = t2
+                elif energies[i + 1] < energies[i] < energies[i - 1]:
+                    tangent = t1
+                else:
+                    demax = max(abs(energies[i + 1] - energies[i]),
+                                abs(energies[i - 1] - energies[i]))
+                    demin = min(abs(energies[i + 1] - energies[i]),
+                                abs(energies[i - 1] - energies[i]))
+                    if energies[i + 1] > energies[i - 1]:
+                        tangent = t2 * demax + t1 * demin
+                    else:
+                        tangent = t2 * demin + t1 * demax
+            # Normalize the tangent vector
+            tangent /= np.linalg.norm(tangent)
             f = forces[i - 1]
             ft = np.vdot(f, tangent)
+            # Remove calculated force parallel to tangent
+            f -= ft * tangent
             if i == imax and self.climb:
-                f -= 2 * ft / tt * tangent
+                # imax not affected by the spring forces. The full force
+                # with component along the elestic band converted
+                # (formula 5 of Paper II)
+                f -= ft * tangent
+            elif self.cornercutting:
+                f1 = -(nt1 - length) * t1 / nt1 * self.k[i - 1]
+                f2 = (nt2 - length) * t2 / nt2 * self.k[i]
+                if (self.climb and abs(i - imax) == 1 and
+                    not np.isnan(energies[i - 1]) and
+                    not np.isnan(energies[i + 1])):
+                    demax = max(abs(energies[i + 1] - energies[i]),
+                                abs(energies[i - 1] - energies[i]))
+                    demin = min(abs(energies[i + 1] - energies[i]),
+                                abs(energies[i - 1] - energies[i]))
+                    f += (f1 + f2) * demin / demax
+                else:
+                    f += f1 + f2
             else:
-                f -= ft / tt * tangent
-                f -= np.vdot(tangent1 * self.k[i - 1] -
-                             tangent2 * self.k[i], tangent) / tt * tangent
-
-            tangent1 = tangent2
+                # Improved parallel spring force (formula 12 of paper I)
+                f += (nt2 * self.k[i] - nt1 * self.k[i - 1]) * tangent
 
         return forces.reshape((-1, 3))
 
@@ -173,11 +253,8 @@ class IDPP(Calculator):
     """Image dependent pair potential.
 
     See:
-
         Improved initial guess for minimum energy path calculations.
-
         Søren Smidstrup, Andreas Pedersen, Kurt Stokbro and Hannes Jónsson
-
         Chem. Phys. 140, 214106 (2014)
     """
 
@@ -454,9 +531,9 @@ class NEBtools:
                      % (Ef, Er, dE))
         return fig
 
-    def get_fmax(self):
+    def get_fmax(self, **kwargs):
         """Returns fmax, as used by optimizers with NEB."""
-        neb = NEB(self._images)
+        neb = NEB(self._images, **kwargs)
         forces = neb.get_forces()
         return np.sqrt((forces**2).sum(axis=1).max())